BBox: Changed the rounding again, the old one fixed Paper Mario but had bugs in Mickey's Magical Mirror. This change fixes the glitches in both games.
Also fixed some compiler warnings.

+++ Source/Core/VideoBackends/D3D/D3DUtil.cpp
@@ -556,7 +556,7 @@ void drawShadedTexQuad(ID3D11ShaderResourceView* texture,
 }
 
 void drawShadedTexSubQuad(ID3D11ShaderResourceView* texture,
-							const MathUtil::Rectangle<float>* rSource,
+							const MathUtil::Rectangle<int>* rSource,
 							int SourceWidth,
 							int SourceHeight,
 							const MathUtil::Rectangle<float>* rDest,
